update posix.fd and posix.file
